+++ core/Lucy/Index/SegWriter.c
@@ -127,7 +127,7 @@ SegWriter_Add_Inverted_Doc_IMP(SegWriter *self, Inverter *inverter,
 static void
 S_adjust_doc_id(SegWriter *self, SegReader *reader, I32Array *doc_map) {
     SegWriterIVARS *const ivars = SegWriter_IVARS(self);
-    size_t doc_count = SegReader_Doc_Max(reader);
+    int32_t doc_count = SegReader_Doc_Max(reader);
     for (size_t i = 1, max = I32Arr_Get_Size(doc_map); i < max; i++) {
         if (I32Arr_Get(doc_map, i) == 0) { doc_count--; }
     }
